agentic_code_review/llm_reviewer/prompts/review_prompts.py
"""Code review prompt templates."""

from langchain_core.prompts import PromptTemplate

# Base template for code review
CODE_REVIEW_TEMPLATE = """You are an expert code reviewer with deep knowledge of software engineering best practices.
Review the following code changes and provide detailed feedback on:

1. Code Quality:
   - Clean code principles
   - Design patterns
   - Code organization
   - Variable/function naming
   - Comments and documentation

2. Performance:
   - Algorithmic efficiency
   - Resource usage
   - Potential bottlenecks

3. Security:
   - Potential vulnerabilities
   - Input validation
   - Authentication/authorization issues
   - Data protection

4. Testing:
   - Test coverage
   - Edge cases
   - Test quality and organization

5. Maintainability:
   - Code duplication
   - Modularity
   - Extensibility
   - Dependencies

Context:
File: {file_path}
Changes:
{code_diff}

Additional Context:
{additional_context}

For each issue found, provide a review comment following this JSON schema:
{format_instructions}

Your response MUST be valid JSON according to the schema above. I will parse your response programmatically.

Notes:
- Location should be in format [file:line] (e.g., [main.py:42])
- Category must be one of: Quality, Performance, Security, Testing, Maintainability
- Severity must be one of: High, Medium, Low
- Description should be clear and specific
- Suggestion should provide actionable improvement steps

Focus on substantive issues that would meaningfully improve the code."""

# Create PromptTemplate instances
code_review_prompt = PromptTemplate(
    input_variables=["file_path", "code_diff", "additional_context", "format_instructions"],
    template=CODE_REVIEW_TEMPLATE,
)

# Template for reviewing test files specifically
TEST_REVIEW_TEMPLATE = """You are an expert in software testing and test-driven development.
Review the following test code changes and provide detailed feedback on:

1. Test Coverage:
   - Test scenarios covered
   - Missing edge cases
   - Integration test aspects

2. Test Quality:
   - Test organization and structure
   - Assertion quality
   - Test isolation
   - Mock/stub usage
   - Test readability

3. Test Maintainability:
   - Test setup and teardown
   - Test data management
   - Test naming conventions
   - Test documentation

Context:
File: {file_path}
Changes:
{code_diff}

Additional Context:
{additional_context}

For each issue found, provide a review comment following this JSON schema:
{format_instructions}

Your response MUST be valid JSON according to the schema above. I will parse your response programmatically.

Notes:
- Location should be in format [file:line] (e.g., [test_main.py:42])
- Category must be one of: Coverage, Quality, Maintainability
- Severity must be one of: High, Medium, Low
- Description should be clear and specific
- Suggestion should provide actionable improvement steps

Focus on improving test reliability, maintainability, and effectiveness."""

# Create PromptTemplate for test review
test_review_prompt = PromptTemplate(
    input_variables=["file_path", "code_diff", "additional_context", "format_instructions"],
    template=TEST_REVIEW_TEMPLATE,
)
